pyproject.toml
[tool.poetry]
name = "voiceassistant"
version = "0.1.0"
description = ""
authors = ["Mohamed Abdelwahab <moh.elwah@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.1.2"
openai = "^1.9.0"
elevenlabs = "^0.2.27"
langchain-openai = "^0.0.3"
python-dotenv = "^1.0.1"
flask = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
